+++ Solver/Integrator/LeeNewmarkFull.cpp
@@ -362,7 +362,7 @@ int LeeNewmarkFull::process_constraint() {
     const auto& D = get_domain();
 
     // process constraint for the first time to obtain proper stiffness
-    if(SUANPAN_SUCCESS != Integrator::process_constraint()) return SUANPAN_FAIL;
+    if(SUANPAN_SUCCESS != LeeNewmarkBase::process_constraint()) return SUANPAN_FAIL;
 
     // this stiffness contains geometry, mass and damping from Newmark::assemble_matrix()
     auto& t_stiff = factory->get_stiffness()->triplet_mat;
@@ -461,7 +461,7 @@ int LeeNewmarkFull::process_constraint() {
         f_mass.get();
 
         // now apply constraints
-        if(SUANPAN_SUCCESS != Integrator::process_constraint()) return SUANPAN_FAIL;
+        if(SUANPAN_SUCCESS != LeeNewmarkBase::process_constraint()) return SUANPAN_FAIL;
         t_stiff.csc_condense();
 
         // move original stiffness matrix back